requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-multipart==0.0.18

httpx==0.28.1
aiohttp==3.11.11
requests==2.32.3
websockets==14.1

sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

influxdb-client==1.48.0

litellm==1.55.8
locust==2.32.4

pandas==2.2.3
numpy==2.2.1

matplotlib==3.10.0
plotly==5.24.1
reportlab==4.2.5
Pillow==11.1.0

pyyaml==6.0.2
click==8.1.8
rich==13.9.4
psutil==6.1.1

watchfiles==1.0.4